Ignore links which are recognized as the catch-all route
