+++ tests/dummy/app/router.js
@@ -11,6 +11,8 @@ Router.map(function() {
     this.route('first');
     this.route('second');
   });
+
+  this.route("404", { path: "*path"});
 });
 
 export default Router;
