+++ tests/dummy/app/templates/application.hbs
@@ -22,4 +22,10 @@
 </div>
 <hr />
 
+normal links:
+<div id="normal-links">
+  [<a href="/reviews">Reviews</a>]
+</div>
+<hr />
+
 {{outlet}}
